requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.13.4
fsspec==2024.3.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
jsonpickle==4.0.2  # Removed the duplicate jsonpickle version 3.0.3
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==24.0
pandas==2.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0  # Added requests
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
pyreadline3==3.5.4
